cribl-utilities-cli/pyproject.toml
[tool.poetry]
name = "cribl-utilities-cli"
version = "0.1.0"
description = ""
authors = ["James Twose <james.twose2711@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
typer = "^0.14.0"
tomli = "^2.2.1"
pydantic = "^2.10.2"
requests = "^2.32.3"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
cribl-utilities-cli = "cribl_utilities_cli.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
